requirements.txt
accelerate==1.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.45.2
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
av==14.0.1
backoff==2.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
braceexpand==0.1.7
cachetools==5.5.1
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.1
cupy-cuda12x==13.3.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.11
decorator==5.1.1
decord==0.6.0
diffusers==0.32.2
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
einx==0.3.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
executing==2.1.0
expelliarmus==1.1.12
fairscale==0.4.4
fastapi==0.115.8
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-genai==1.2.0
gradio==5.16.0
gradio_client==1.7.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
hydra-core==1.3.2
identify==2.6.7
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importRosbag==1.0.4
iopath==0.1.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.17
kiwisolver==1.4.8
kornia==0.7.4
kornia_rs==0.1.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.11.9
llvmlite==0.43.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
momo-opt==0.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
narwhals==1.27.1
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nir==1.0.4
nirtorch==1.0
nltk==3.9.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.63.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pbr==6.1.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pooch==1.8.2
portalocker==3.1.1
pre_commit==4.1.0
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-multipart==0.0.20
python-slugify==8.0.4
pytorch-ignite==0.5.1
pytorch-lightning==2.5.0.post0
pytorchvideo==0.1.5
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rarfile==4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.9.6
safehttpx==0.1.6
safetensors==0.5.0
salesforce-lavis==1.0.2
scikit-image==0.25.1
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.4.1
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
snntorch==0.9.1
soundfile==0.13.0
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spikingjelly==0.0.0.0.14
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
streamlit==1.42.0
sympy==1.13.1
tabulate==0.9.0
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2025.1.10
tiktoken==0.3.3
timm==1.0.13
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
tonic==1.5.0
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.6.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.2.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
vector-quantize-pytorch==1.21.1
virtualenv==20.29.2
wandb==0.19.1
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.111
webencodings==0.5.1
websockets==14.2
wrapt==1.17.2
yacs==0.1.8
yarl==1.18.3
zipp==3.21.0
